refactoring components
