+++ client/src/components/Modal/CartModal/styles.js
@@ -1,0 +1,26 @@
+export const style = {
+	modal: {
+		position: 'absolute',
+		overflowY: 'scroll',
+		top: '50%',
+		left: '50%',
+		width: 320,
+		boxSizing: 'border-box',
+		transform: 'translate(-50%, -50%)',
+		bgcolor: '#fff',
+		padding: '15px 10px 0',
+		maxHeight: '80vh',
+	},
+	link: {
+		textDecoration: 'none',
+		display: 'block',
+		width: '100%',
+		margin: '1.5rem 0'
+	},
+	button: {
+		width: '100%'
+	},
+	title: {
+		margin: '1rem 0'
+	}
+}
