+++ client/src/components/Modal/LoginModal/LoginModal.js
@@ -1,0 +1,15 @@
+import React from 'react'
+import Form from '../../Form/Form'
+import BasicModal from '../Modal'
+import { style } from './styles'
+
+const LoginModal = () => {
+	return (
+		<BasicModal
+			style={style}
+			body={<Form />}
+		/>
+	)
+}
+
+export default LoginModal
