+++ client/src/components/Modal/LoginModal/styles.js
@@ -1,0 +1,11 @@
+export const style = {
+	position: 'absolute',
+	top: '50%',
+	left: '50%',
+	transform: 'translate(-50%, -50%)',
+	bgcolor: '#fff',
+	padding: '20px',
+	['@media (max-width:440px)']: {
+		padding: '5px'
+	}
+}
