+++ client/src/components/NavBar/NavBarIcons/CartIcon/CartIcon.jsx
@@ -5,7 +5,7 @@ import LocalMallIcon from '@mui/icons-material/LocalMall'
 import { useStyles } from './styles'
 import { useSelector, useDispatch } from 'react-redux'
 import modalActions from '../../../../store/Modal'
-import CartModal from '../../../Modal/CartModal'
+import CartModal from '../../../Modal/CartModal/CartModal'
 import { shoppingBagSelectors } from '../../../../store/ShoppingBag'
 import { Box } from '@mui/system'
