+++ client/src/pages/Main/Main.js
@@ -5,7 +5,7 @@ import Popular from '../../components/Popular/Popular'
 import Carousel from '../../components/Carousel/Carousel'
 import { slidesOperations, slidesSelectors } from '../../store/Slider'
 import { useDispatch, useSelector } from 'react-redux'
-import UseCeo from '../../utils/customHooks/useCeo'
+import UseSeo from '../../utils/customHooks/useSeo'
 
 const Main = () => {
 	const slides = useSelector(slidesSelectors.getSlides())
@@ -17,7 +17,7 @@ const Main = () => {
 
 	return (	
 		<>
-			<UseCeo 
+			<UseSeo 
 				title = {'Bedding Shop'}
 				description = {'Bedding Shop, all for '}
 				keywords = {'bedroom, Kitchen, Loungewear, bedding store, luxury bedding sets, king size bedspreads, king bed sheets, comforters on sale'}
