+++ client/src/pages/Page404/Page404.js
@@ -2,25 +2,10 @@ import React from 'react'
 import Grid from '@mui/material/Grid'
 import Typography from '@mui/material/Typography'
 import Button from '@mui/material/Button'
-import {ThemeProvider, styled} from '@mui/material/styles'
+import { ThemeProvider } from '@mui/material/styles'
 import theme from '../../utils/Theme'
 import { Link as RouterLink } from 'react-router-dom'
-
-
-const StyledGrid = styled(Grid)(() => ({
-	backgroundImage: 'url(https://telegra.ph/file/a833f6b2d325d08432a9e.jpg)',
-	backgroundSize: 'cover',
-	backgroundPosition: 'center',
-	paddingBottom: '0.67em',
-	alignItems:'center',
-	objectFit: 'cover',
-}))
-
-
-const styles = {
-	padding: '72px 50px 0px 50px', textAlign: 'center',
-}
-
+import { StyledGrid, styles } from './styles'
 
 const Error404 = () => {
 	return (
@@ -30,18 +15,18 @@ const Error404 = () => {
 		>
 			<Grid item>
 				<Typography variant={'h1'}>
-    404
+					404
 				</Typography>
 			</Grid>
 			<Grid item>
-				<Typography style={styles}  variant={'h2'}>
-      Oops. Page not found
+				<Typography style={styles} variant={'h2'}>
+					Oops. Page not found
 				</Typography>
 			</Grid>
 			<Grid item>
 				<Typography style={styles} variant={'body1'}>
-      You may have mistyped the
-      address or the page may have moved.
+					You may have mistyped the
+					address or the page may have moved.
 				</Typography>
 			</Grid>
 			<Grid style={styles} item>
