+++ client/src/pages/Page404/styles.js
@@ -1,0 +1,15 @@
+import Grid from '@mui/material/Grid'
+import { styled } from '@mui/material/styles'
+
+export const StyledGrid = styled(Grid)(() => ({
+	backgroundImage: 'url(https://telegra.ph/file/a833f6b2d325d08432a9e.jpg)',
+	backgroundSize: 'cover',
+	backgroundPosition: 'center',
+	paddingBottom: '0.67em',
+	alignItems: 'center',
+	objectFit: 'cover',
+}))
+
+export const styles = {
+	padding: '72px 50px 0px 50px', textAlign: 'center',
+}
