+++ client/src/pages/ProductDetails/ProductDetails.js
@@ -2,13 +2,14 @@ import React, { useEffect } from 'react'
 import { useParams } from 'react-router-dom'
 import { useDispatch, useSelector } from 'react-redux'
 import { ProductOperations, ProductSelector } from '../../store/Product'
-import { Container, Grid, Box, Typography, styled } from '@mui/material'
+import { Container, Grid, Box } from '@mui/material'
 import ProductDescription from '../../components/ProductDescription/ProductDescription'
 import Carousel from '../../components/Carousel/Carousel'
 import RelatedItemsList from '../../components/RelatedItems/RelatedItemsList'
 import BackdropLoader from '../../components/UI/BackdropLoader/BackdropLoader'
-import UseCeo from '../../utils/customHooks/useCeo'
+import UseSeo from '../../utils/customHooks/useSeo'
 import { favoritesOperations } from '../../store/Favorites'
+import { StyledTypography } from './styles'
 
 const ProductDetails = () => {
 	const { id } = useParams()
@@ -18,14 +19,6 @@ const ProductDetails = () => {
 	const parent = useSelector(ProductSelector.getParent())
 	const favorites = JSON.parse(localStorage.getItem('favorites')) || []
 
-	const StyledTypography = styled(Typography)(() => ({
-		fontSize: '32px',
-		textTransform: 'uppercase',
-		fontWeight: 400,
-		margin: 'auto',
-		alignSelf: 'center',
-	}))
-	
 	useEffect(() => {
 		favoritesOperations.fetchFavorites(favorites)(dispatch)
 		// eslint-disable-next-line react-hooks/exhaustive-deps
@@ -49,18 +42,18 @@ const ProductDetails = () => {
 	}, [activeProduct])
 
 	if (isLoading) {
-		return <BackdropLoader open={isLoading}/>
+		return <BackdropLoader open={isLoading} />
 	}
 
 	return (
 		<>
-			<UseCeo 
-				title = {parent ? parent.name : 'Product Details'}
-				description = {parent ? parent.description : 'Product details'}
-				keywords = {
-					parent 
+			<UseSeo
+				title={parent ? parent.name : 'Product Details'}
+				description={parent ? parent.description : 'Product details'}
+				keywords={
+					parent
 						? `${parent.name}, ${parent.manufacturer},  ${parent.brand}, 
-						${parent.seller}, ${parent.manufacturerCountry}` 
+						${parent.seller}, ${parent.manufacturerCountry}`
 						: null
 				}
 			/>
@@ -70,14 +63,14 @@ const ProductDetails = () => {
 				)}
 				{activeProduct && <><Grid container spacing={4}>
 					<Grid item md={6} xs={12}>
-						<Carousel slides={activeProduct.imageUrls} product={true}/>
+						<Carousel slides={activeProduct.imageUrls} product={true} />
 					</Grid>
 					<Grid item md={6} xs={12}>
-						<ProductDescription/>
+						<ProductDescription />
 					</Grid>
 				</Grid>
-				<Box style={{marginTop: '80px'}}>
-					<RelatedItemsList/>
+				<Box style={{ marginTop: '80px' }}>
+					<RelatedItemsList />
 				</Box>
 				</>
 				}
