+++ client/src/routes/AppRoutes.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Route, Switch} from 'react-router-dom'
+import { Route, Switch } from 'react-router-dom'
 import ProtectedRoute from '../components/ProtectedRoute/ProtectedRoute'
 import Main from '../pages/Main/Main'
 import ProductDetails from '../pages/ProductDetails/ProductDetails'
@@ -21,15 +21,12 @@ const AppRoutes = () => {
 			<ProtectedRoute exact path='/favorites'><Favorites /></ProtectedRoute>
 			<Route exact path='/product-details/:id'><ProductDetails /></Route>
 			<Route exact path='/cart'><Cart /></Route>
-
-			<Route exact path='/shop/catalog'><Catalog/></Route>
+			<Route exact path='/shop/catalog'><Catalog /></Route>
 			<Route exact path='/search'><Search /></Route>
 			<Route exact path='/info/:customId'><StaticPages /></Route>
-
 			<Route exact path='/subscription/:email'><HandleSubscribe /></Route>
-			<ProtectedRoute exact path='/checkout'><Checkout/></ProtectedRoute>
-			<ProtectedRoute exact path='/user-profile'><UserProfilePage/></ProtectedRoute>
-			{/*<ProtectedRoute exact path='/complete-order'><CompletePay /></ProtectedRoute>*/}
+			<ProtectedRoute exact path='/checkout'><Checkout /></ProtectedRoute>
+			<ProtectedRoute exact path='/user-profile'><UserProfilePage /></ProtectedRoute>
 			<Route exact path='/complete-order'><CompletePay /></Route>
 			<Route exact path='*'><Page404 /></Route>
 		</Switch>
